Latest commit

+++ lab4_sistadelen/Operation.java
@@ -1,9 +1,9 @@
 package lab4_sistadelen;
 
 public interface Operation {
-
+	
 	public char symbol();
 	public int width(int rows, int cols);
-	public int evaluate(int a, int b);
+	public double evaluate(int a, int b);
 	
 }
